+++ app/elasticsearch/CirceDecoders.scala
@@ -1,4 +1,4 @@
-package es
+package elasticsearch
 
 import java.time.OffsetDateTime
