+++ app/views/ViewHelper.scala
@@ -3,7 +3,7 @@ package views
 import java.time.{OffsetDateTime, ZoneOffset}
 import java.time.format.DateTimeFormatter
 
-import models.ApiKey
+import apikeys.model
 
 object ViewHelper {
 
@@ -12,7 +12,7 @@ object ViewHelper {
       .atZoneSameInstant(ZoneOffset.UTC)
       .format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")) + " GMT"
 
-  def statusBadge(apiKey: ApiKey) =
+  def statusBadge(apiKey: model) =
     if (apiKey.active)
       <span class="label label-primary">Active</span>
     else
