+++ app/views/deployments/search.scala.html
@@ -1,4 +1,4 @@
-@(page: es.ES.Page[Identified[Deployment]], team: Option[String], service: Option[String], buildId: Option[String], showAdminColumn: Boolean)(implicit user: com.gu.googleauth.UserIdentity, request: RequestHeader)
+@(page: elasticsearch.Elastic55.Page[Identified[Deployment]], team: Option[String], service: Option[String], buildId: Option[String], showAdminColumn: Boolean)(implicit user: com.gu.googleauth.UserIdentity, request: RequestHeader)
 
 @import views.ViewHelper._
 @import helper._
